% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boston.R
\docType{data}
\name{Boston}
\alias{Boston}
\title{Boston housing data}
\format{
A data frame with 506 rows and 14 variables:
\itemize{
\item CRIM - per capita crime rate by town
\item ZN - proportion of residential land zoned for lots over 25,000 sq.ft.
\item INDUS - proportion of non-retail business acres per town.
\item CHAS - Charles River dummy variable (1 if tract bounds river; 0 otherwise)
\item NOX - nitric oxides concentration (parts per 10 million)
\item RM - average number of rooms per dwelling
\item AGE - proportion of owner-occupied units built prior to 1940
\item DIS - weighted distances to five Boston employment centres
\item TAX - full-value property-tax rate per $10,000
\item RAD - index of accessibility to radial highways
\item PTRATIO - pupil-teacher ratio by town
\item B - 1000(Bk - 0.63)^2 where Bk is the proportion of blacks by town
\item LSTAT - percentage lower status of the population
\item MEDV - Median value of owner-occupied homes in $1000's
}
}
\source{
\url{http://lib.stat.cmu.edu/datasets/boston}
}
\usage{
Boston
}
\description{
A dataset containing information collected by the U.S Census Service concerning housing
in the area of Boston, originally published by Harrison and Rubinfeld (1978).
}
\references{
Harrison, D and Rubinfeld, DL (1978)
`Hedonic prices and the demand for clean air',
Journal of Environmental Economics and Management 5, 81-102.
}
\keyword{datasets}
