% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphhelpfns.R
\name{dag2adjacencymatrix}
\alias{dag2adjacencymatrix}
\title{Deriving an adjacency matrix of a graph}
\usage{
dag2adjacencymatrix(g)
}
\arguments{
\item{g}{graph, object of class \code{\link[graph]{graphNEL}} (package `graph')}
}
\value{
a square matrix whose dimensions are the number of nodes in the graph g, where element
\code{[i,j]} equals \code{1} if there is a directed edge from node \code{i} to node \code{j} in the graph \code{g},
 and \code{0} otherwise
}
\description{
This function derives the adjacency matrix corresponding to a graph object
}
