% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNfns.R
\name{full2compact}
\alias{full2compact}
\title{Deriving a compact adjacency matrix of a DBN}
\usage{
full2compact(DBNmat, b = 0)
}
\arguments{
\item{DBNmat}{a square matrix, representing the structure of an unrolled DBN; the size of matrix is slices*dyn+b; all static variables are assumed to be in the first b rows and columns of the matrix}

\item{b}{integer, number of static variables; 0 by default}
}
\description{
This function transforms an unrolled adjacency matrix of DBN into a compact representation
}
\examples{
full2compact(DBNunrolled,b=3)
}
