% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphhelpfns.R
\name{getSubGraph}
\alias{getSubGraph}
\title{Deriving subgraph}
\usage{
getSubGraph(adj, nodes)
}
\arguments{
\item{adj}{square adjacency matrix with elements in \code{\{0,1\}}, representing a graph}

\item{nodes}{vector of node names of the subgraph; should be a subset of column names of 'adj'}
}
\value{
adjacency matrix of a subgraph which includes all 'nodes'
}
\description{
This function derives an adjacency matrix of a subgraph based on the adjacency matrix of a full graph and a list of nodes
}
\examples{
getSubGraph(Asiamat,c("E","B","D","X"))
}
