% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsim.R
\docType{data}
\name{gsim}
\alias{gsim}
\title{A simulated data set from a Gaussian continuous Bayesian network}
\format{
A data frame with 1000 rows representing observations of 100 continuous variables: V1, ..., V100
}
\usage{
gsim
}
\description{
A synthetic dataset containing 1000 observations generated from a random DAG with 100 continuous nodes.
}
\keyword{datasets}
