% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsim100.R
\docType{data}
\name{gsim100}
\alias{gsim100}
\title{A simulated data set from a Gaussian continuous Bayesian network}
\format{
A data frame with 100 rows representing observations of 100 continuous variables: V1, ..., V100
}
\usage{
gsim100
}
\description{
A synthetic dataset containing 100 observations generated from a random DAG with 100 continuous nodes.
}
\keyword{datasets}
