% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsimadj.R
\docType{data}
\name{gsimmat}
\alias{gsimmat}
\title{An adjacency matrix of a simulated dataset}
\format{
A binary matrix with 100 rows and 100 columns representing an adjacency matrix of a DAG with 100 nodes: V1, ..., V100
}
\usage{
gsimmat
}
\description{
An adjacency matrix representing the ground truth DAG used to generate a synthetic dataset with observations of 100 continuous variables.
}
\keyword{datasets}
