% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSTRING.R
\docType{data}
\name{interactions}
\alias{interactions}
\title{interactions dataset}
\format{
A data frame with 179 rows and 3 columns; 
\itemize{
\item node1 character, name of a gene 
\item node2 character, name of a gene 
\item combined_score interaction score, reflecting confidence in the fact that interaction between gene1 and gene2 is possible
}
each row represents a possible interaction between two genes
}
\source{
\url{https://string-db.org/}
}
\usage{
interactions
}
\description{
A data frame containing possible interactions between genes from \code{kirp} and \code{kirc}
data sets
}
\keyword{datasets}
