% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSTRING.R
\docType{data}
\name{intSTRING}
\alias{intSTRING}
\title{intSTRING dataset}
\format{
A data frame with 179 rows and 2 columns; each row represent a possible interaction between two genes
}
\source{
\url{https://string-db.org/}
}
\usage{
intSTRING
}
\description{
A data frame containing possible interactions between genes from \code{kirp} and \code{kirc}
data sets
}
\keyword{datasets}
