% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KIRP.R
\docType{data}
\name{kirp}
\alias{kirp}
\title{KIRP dataset}
\format{
An object of class \code{matrix} with 161 rows and 40 columns.
}
\usage{
kirp
}
\description{
Mutation data from TCGA kidney renal papillary cell cohort (KIRP). 
Mutations are picked according to q-value computed 
by MutSigCV (q<0.1) or connected in networks discovered by Kuipers et al. 2018.
}
\details{
{
Each variable represents a gene. If in sample i gene j contains a mutation, than j-th element in 
row i equals 1, and 0 otherwise.
The rows are named according to sample names in TCGA.
The columns are named according to gene symbols.
}
}
\references{
\url{https://portal.gdc.cancer.gov/}

\url{http://firebrowse.org/iCoMut/?cohort=kich}

Lawrence, M. et al. Mutational heterogeneity in cancer and the search for new cancer-associated genes. Nature 499, 214-218 (2013)
}
\keyword{datasets}
