% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSTRING.R
\docType{data}
\name{mapSTRING}
\alias{mapSTRING}
\title{mapSTRING dataset}
\format{
A data frame with 46 rows and two columns:
\itemize{
\item queryItem character, name used for structure learning (for example see column names of \code{\link{kirp}})
\item preferredName character, name used in STRING interactions data set (see \code{\link{intSTRING}})
}
}
\source{
\url{https://string-db.org/}
}
\usage{
mapSTRING
}
\description{
A data frame containing mapping between names of genes used in \code{kirp}/\code{kirc}
data sets and names used in STRING interactions list (see \code{\link{intSTRING}}).
}
\keyword{datasets}
