% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotS3.R
\name{plot.MCMCmult}
\alias{plot.MCMCmult}
\title{S3 methods for class 'MCMCmult'}
\usage{
\method{plot}{MCMCmult}(x, ...)
}
\arguments{
\item{x}{object of class 'MCMCmult'}

\item{...}{ignored}
}
\description{
Plotting object of class 'MCMCmult'
}
