% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotS3.R
\name{plot.MCMCres}
\alias{plot.MCMCres}
\title{S3 methods for class 'MCMCres'}
\usage{
\method{plot}{MCMCres}(x, ...)
}
\arguments{
\item{x}{object of class 'MCMCres'}

\item{...}{ignored}
}
\description{
Plotting object of class 'MCMCres'
}
