% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printmethods.R
\name{print.MCMCmult}
\alias{print.MCMCmult}
\title{S3 methods for class 'MCMCmult'}
\usage{
\method{print}{MCMCmult}(x, ...)
}
\arguments{
\item{x}{object of class 'MCMCmult'}

\item{...}{ignored}
}
\description{
Prints object of class 'MCMCmult'
}
