% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printmethods.R
\name{print.MCMCres}
\alias{print.MCMCres}
\title{S3 methods for class 'MCMCres'}
\usage{
\method{print}{MCMCres}(x, ...)
}
\arguments{
\item{x}{object of class 'MCMCres'}

\item{...}{ignored}
}
\description{
Prints object of class 'MCMCres'
}
