% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printmethods.R
\name{print.MCMCscoretab}
\alias{print.MCMCscoretab}
\title{S3 methods for class 'MCMCscoretab'}
\usage{
\method{print}{MCMCscoretab}(x, ...)
}
\arguments{
\item{x}{object of class 'MCMCscoretab'}

\item{...}{ignored}
}
\description{
Prints object of class 'MCMCscoretab'
}
