% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printmethods.R
\name{print.scoreparameters}
\alias{print.scoreparameters}
\title{S3 methods for class 'scoreparameters'}
\usage{
\method{print}{scoreparameters}(x, ...)
}
\arguments{
\item{x}{object of class 'scoreparameters'}

\item{...}{ignored}
}
\description{
Print object of class 'scoreparameters'
}
