% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreagainstDBN.R
\name{scoreagainstDBN}
\alias{scoreagainstDBN}
\title{Score against DBN}
\usage{
scoreagainstDBN(
  scorepar,
  incidence,
  datatoscore = NULL,
  marginalise = FALSE,
  onlymain = FALSE,
  datainit = NULL
)
}
\arguments{
\item{scorepar}{object of class 'scoreparameters'}

\item{incidence}{adjacency matrix of a DAG}

\item{datatoscore}{matrix or vector containing observations to be scored}

\item{marginalise}{(logical) should marginal score be used?}

\item{onlymain}{(logical) should static nodes be included in the score?}

\item{datainit}{optional, in case of unbalanced design, the mean score of available samples for T0 are computed}
}
\value{
vector of log-scores
}
\description{
Scoring observations against DBN structures
}
\author{
Polina Suter
}
