% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.MCMCmult}
\alias{summary.MCMCmult}
\title{S3 methods for class 'MCMCmult'}
\usage{
\method{summary}{MCMCmult}(object, ...)
}
\arguments{
\item{object}{object of class 'MCMCmult'}

\item{...}{ignored}
}
\description{
Summary of object of class 'MCMCmult'
}
