% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.MCMCres}
\alias{summary.MCMCres}
\title{S3 methods for class 'MCMCres'}
\usage{
\method{summary}{MCMCres}(object, ...)
}
\arguments{
\item{object}{object of class 'MCMCres'}

\item{...}{ignored}
}
\description{
Summary of object of class 'MCMCres'
}
