% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.MCMCscoretab}
\alias{summary.MCMCscoretab}
\title{S3 methods for class 'MCMCscoretab'}
\usage{
\method{summary}{MCMCscoretab}(object, ...)
}
\arguments{
\item{object}{object of class 'MCMCscoretab'}

\item{...}{ignored}
}
\description{
Summary of object of class 'MCMCscoretab'
}
