% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.scoreparameters}
\alias{summary.scoreparameters}
\title{S3 methods for class 'scoreparameters'}
\usage{
\method{summary}{scoreparameters}(object, ...)
}
\arguments{
\item{object}{object of class 'scoreparameters'}

\item{...}{ignored}
}
\description{
Summary of object of class 'scoreparameters'
}
