% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNfns.R
\name{full2compact}
\alias{full2compact}
\title{Deriving a compact adjacency matrix of a DBN}
\usage{
full2compact(DBNmat, n.dynamic, n.static = 0)
}
\arguments{
\item{DBNmat}{a square matrix, representing the structure of an unrolled DBN; the size of matrix is n.slices*n.dynamic+n.static; all static variables are assumed to be in the first n.static rows and columns of the matrix}

\item{n.dynamic}{integer, number of dynamic variables in each time slice}

\item{n.static}{integer, number of static variables}
}
\description{
This function transforms an unrolled adjacency matrix of DBN into a compact representation
}
\examples{
full2compact(DBNunrolled,n.dynamic=12,n.static=3)
}
